refactor(search-view): extract useRelayStatus hook
